corrections to functions and stability
