+++ Client/src/app/app.component.ts
@@ -9,6 +9,7 @@ import { TabsPage } from '../pages/tabs/tabs';
   templateUrl: 'app.html'
 })
 export class MyApp {
+  
   rootPage = TabsPage;
 
   constructor(platform: Platform) {
