+++ Client/src/pages/schedules/schedules.html
@@ -1,0 +1,72 @@
+<ion-header>
+  <ion-navbar color="primary">
+    <ion-title>
+      Events
+    </ion-title>
+  </ion-navbar>
+</ion-header>
+
+<ion-content>
+<ion-list no-lines>
+  <ion-item-divider color="light">{{date | date:'mediumDate'}}</ion-item-divider>
+  <ion-item [ngClass]="{ spacer: spacer }">
+      <ion-icon id="event1" name="ios-calendar-outline" item-left></ion-icon>
+      <h2>Sprint Meeting</h2>
+      <p>9:15 AM - 10:45 AM</p>
+      <ion-icon (click)="openModal({charNum: 0})" name="ios-arrow-dropright-outline" item-right></ion-icon>
+  </ion-item>
+  <ion-item [ngClass]="{ spacer: spacer }">
+      <ion-icon id="event2" name="ios-calendar-outline" item-left></ion-icon>
+      <h2>Digital Marketing</h2>
+      <p>11:30 AM - 14:00 PM</p>
+      <ion-icon (click)="openModal({charNum: 1})" name="ios-arrow-dropright-outline" item-right></ion-icon>
+  </ion-item>
+
+  <ion-item-divider color="light">{{tomorrow | date:'mediumDate'}}</ion-item-divider>
+  <ion-item [ngClass]="{ spacer: spacer }">
+      <ion-icon id="event1" name="ios-calendar-outline" item-left></ion-icon>
+      <h2>Sprint Meeting</h2>
+      <p>9:15 AM - 10:45 AM</p>
+      <ion-icon (click)="openModal({charNum: 2})" name="ios-arrow-dropright-outline" item-right></ion-icon>
+  </ion-item>
+  <ion-item [ngClass]="{ spacer: spacer }">
+      <ion-icon name="ios-calendar-outline" item-left></ion-icon>
+      <h2>Product Presentation</h2>
+      <p>12:15 PM - 16:35 PM</p>
+      <ion-icon (click)="openModal({charNum: 3})" name="ios-arrow-dropright-outline" item-right></ion-icon>
+  </ion-item>
+  <ion-item [ngClass]="{ spacer: spacer }">
+      <ion-icon name="ios-calendar-outline" item-left></ion-icon>
+      <h2>Team Meeting</h2>
+      <p>18:15 PM - 18:45 PM</p>
+      <ion-icon (click)="openModal({charNum: 4})" name="ios-arrow-dropright-outline" item-right></ion-icon>
+  </ion-item>
+  <ion-item [ngClass]="{ spacer: spacer }">
+      <ion-icon name="ios-calendar-outline" item-left></ion-icon>
+      <h2>Job Conference</h2>
+      <p>19:45 PM - 20:45 PM</p>
+      <ion-icon (click)="openModal({charNum: 5})" name="ios-arrow-dropright-outline" item-right></ion-icon>
+  </ion-item>
+
+  <ion-item-divider color="light">{{dayaftertomorrow | date:'mediumDate'}}</ion-item-divider>
+  <ion-item [ngClass]="{ spacer: spacer }">
+      <ion-icon id="event1" name="ios-calendar-outline" item-left></ion-icon>
+      <h2>Sprint Meeting</h2>
+      <p>9:15 AM - 10:45 AM</p>
+      <ion-icon (click)="openModal({charNum: 6})" name="ios-arrow-dropright-outline" item-right></ion-icon>
+  </ion-item>
+  <ion-item [ngClass]="{ spacer: spacer }">
+      <ion-icon name="ios-calendar-outline" item-left></ion-icon>
+      <h2>Business Meeting</h2>
+      <p>13:45 PM - 15:45 PM</p>
+      <ion-icon (click)="openModal({charNum: 7})" name="ios-arrow-dropright-outline" item-right></ion-icon>
+  </ion-item>
+   <ion-item [ngClass]="{ spacer: spacer }">
+      <ion-icon name="ios-calendar-outline" item-left></ion-icon>
+      <h2>Product Presentation</h2>
+      <p>17:15 PM - 20:45 PM</p>
+      <ion-icon (click)="openModal({charNum: 8})" name="ios-arrow-dropright-outline" item-right></ion-icon>
+  </ion-item>
+</ion-list>
+
+</ion-content>
