--- Client/src/pages/shedules/shedules.scss
@@ -1,17 +1,0 @@
-page-events {
-
-.item-divider-md{
-    font-family: Arial, Helvetica, sans-serif;
-    font-size: 1.15em;
-}
-
-#event1{
-    color: green;
-}
-
-#event2{
-    color: blue;
-}
-
-
-}
