+++ Client/src/pages/speakers/speakers.html
@@ -18,7 +18,7 @@
 
     <ion-item>
       <ion-icon name="person" item-left large></ion-icon>
-      <h2>Jordan Belfort</h2>
+      <h2>Doris Greene</h2>
     </ion-item>
 
     <ion-item>
@@ -113,7 +113,7 @@
 
     <ion-item>
       <ion-icon name="person" item-left large></ion-icon>
-      <h2>Jordan Belfort</h2>
+      <h2>Sage Rodriguez</h2>
     </ion-item>
 
     <ion-item>
@@ -159,7 +159,7 @@
 
     <ion-item>
       <ion-icon name="person" item-left large></ion-icon>
-      <h2>Jordan Belfort</h2>
+      <h2>Philip Chaney</h2>
     </ion-item>
 
     <ion-item>
@@ -205,7 +205,7 @@
 
     <ion-item>
       <ion-icon name="person" item-left large></ion-icon>
-      <h2>Jordan Belfort</h2>
+      <h2>Mason Porter</h2>
     </ion-item>
 
     <ion-item>
