+++ Client/src/pages/tabs/tabs.html
@@ -1,7 +1,6 @@
 <ion-tabs>
   <ion-tab [root]="tab1Root" tabTitle="Home" tabIcon="home"></ion-tab>
   <ion-tab [root]="tab2Root" tabTitle="Speakers" tabIcon="md-people"></ion-tab>
-  <ion-tab [root]="tab3Root" tabTitle="Chat" tabIcon="md-chatbubbles"></ion-tab>
-  <ion-tab [root]="tab4Root" tabTitle="Map" tabIcon="map"></ion-tab>
-  <ion-tab [root]="tab5Root" tabTitle="Shedules" tabIcon="calendar"></ion-tab>
+  <ion-tab [root]="tab3Root" tabTitle="Map" tabIcon="map"></ion-tab>
+  <ion-tab [root]="tab4Root" tabTitle="Events" tabIcon="calendar"></ion-tab>
 </ion-tabs>
