+++ Client/src/pages/tabs/tabs.ts
@@ -2,9 +2,8 @@ import { Component } from '@angular/core';
 // Importing the tab pages.
 import { HomePage } from '../home/home';
 import { SpeakersPage } from '../speakers/speakers';
-import { ChatsPage } from '../chat/chat';
 import { MapPage } from '../map/map';
-import { ShedulesPage } from '../shedules/shedules';
+import { ShedulesPage } from '../schedules/schedules';
 
 @Component({
   templateUrl: 'tabs.html'
@@ -14,9 +13,8 @@ export class TabsPage {
   // should be each tab's root Page
   tab1Root: any = HomePage;
   tab2Root: any = SpeakersPage;
-  tab3Root: any = ChatsPage;
-  tab4Root: any = MapPage;
-  tab5Root: any = ShedulesPage;
+  tab3Root: any = MapPage;
+  tab4Root: any = ShedulesPage;
 
   constructor() {
